logic_worker_edrian/app/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
deepseek
httpx==0.27.2
requests==2.31.0
PyJWT==2.8.0
six
typing
pydantic